REFACTOR: Updated scheduling to use cron
-Cron gives more precise control over when to run real-time market update jobs
