--- src/main/java/edu/bu/cs673/stockportfolio/service/portfolio/scheduledjobs/JobConfiguration.java
@@ -1,6 +1,0 @@
-package edu.bu.cs673.stockportfolio.service.portfolio.scheduledjobs;
-
-public interface JobConfiguration {
-
-    void scheduleByFixedRate();
-}
